requirements.txt
djongo==1.3.0
Django==2.2.5
djangorestframework==3.11.0
